Separate testing of utility function for logging out to another file
